mcp_servers/telegram/requirements.txt
mcp
uvicorn
httpx
python-dotenv
